requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
httpx>=0.25.0
pyyaml>=6.0
aiosqlite>=0.19.0
playwright>=1.40.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
ddgs>=6.0
Pillow>=10.0
python-multipart>=0.0.6
pandas>=2.1.0
matplotlib>=3.8.0
openpyxl>=3.1.0
PyMuPDF>=1.23.0
yt-dlp>=2024.1.0
